+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 (!) - breaking change
 (#xyz) - fixed in the given PR
 
-## Unreleased
+## [0.8.0] - 2026-08-01
 ### Added
 - `Image::width()` and `Image::height()`
 - `Image::save()`